refactor: remove unused formFields struct and related utility functions to simplify form handling
